+++ src/com/marketplace/dal/partner/PartnerDAL.java
@@ -1,4 +1,4 @@
-package com.marketplace.dal;
+package com.marketplace.dal.partner;
 
 import java.util.List;
 
@@ -8,8 +8,8 @@ import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
 import org.hibernate.cfg.Configuration;
 
-import com.marketplace.domain.Customer;
-import com.marketplace.domain.Partner;
+import com.marketplace.domain.customer.Customer;
+import com.marketplace.domain.partner.Partner;
 
 public class PartnerDAL {
