+++ src/com/marketplace/dal/payment/PaymentDAL.java
@@ -1,4 +1,4 @@
-package com.marketplace.dal;
+package com.marketplace.dal.payment;
 
 import java.util.List;
 import java.util.ArrayList;
@@ -6,9 +6,9 @@ import java.util.ArrayList;
 import org.hibernate.*;
 import org.hibernate.cfg.Configuration;
 
-import com.marketplace.domain.Payment;
-import com.marketplace.domain.Address;
-import com.marketplace.domain.Customer;
+import com.marketplace.domain.address.Address;
+import com.marketplace.domain.customer.Customer;
+import com.marketplace.domain.payment.Payment;
 
 public class PaymentDAL {
