+++ src/com/marketplace/domain/customer/Customer.java
@@ -1,4 +1,4 @@
-package com.marketplace.domain;
+package com.marketplace.domain.customer;
 
 public class Customer {
